+++ History.md
@@ -3,8 +3,10 @@
 
  * add plates support
  * add teacup support
+ * add liquid-node support
  * allow absolute and relative partial paths
  * extend dot options
+ * support layouts in vash
 
 0.14.0 / 2016-01-24
 ===================